refactor: share token count utilities
